pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "payment-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.120.2",
    "fastmcp>=2.13.0.2",
    "python-dotenv>=1.0",
    "requests>=2.32.5",
    "scalekit-sdk-python>=2.4.10",
    "uvicorn>=0.30.6",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
